// src/services/rest-api/customer-service/model/combo-box.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Customer Service API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.0.242
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface ComboBox
 */
export interface ComboBox {
    /**
     * 
     * @type {string}
     * @memberof ComboBox
     */
    'rValue'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof ComboBox
     */
    'rText'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof ComboBox
     */
    'rSort'?: string | null;
}