// src/services/rest-api/customer-service/model/occupation-info-model.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Customer Service API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.0.267
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface OccupationInfoModel
 */
export interface OccupationInfoModel {
    /**
     * 
     * @type {string}
     * @memberof OccupationInfoModel
     */
    'occupationCode'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof OccupationInfoModel
     */
    'jobWorkPlace'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof OccupationInfoModel
     */
    'jobPosition'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof OccupationInfoModel
     */
    'jobDepartment'?: string | null;
}