// src/services/rest-api/customer-service/model/occupation-info-response-data-response.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Customer Service API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.0.267
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import { OccupationInfoResponse } from './occupation-info-response';

/**
 * 
 * @export
 * @interface OccupationInfoResponseDataResponse
 */
export interface OccupationInfoResponseDataResponse {
    /**
     * 
     * @type {number}
     * @memberof OccupationInfoResponseDataResponse
     */
    'status'?: number;
    /**
     * 
     * @type {{ [key: string]: Array<string>; }}
     * @memberof OccupationInfoResponseDataResponse
     */
    'errors'?: { [key: string]: Array<string>; } | null;
    /**
     * 
     * @type {string}
     * @memberof OccupationInfoResponseDataResponse
     */
    'message'?: string | null;
    /**
     * 
     * @type {OccupationInfoResponse}
     * @memberof OccupationInfoResponseDataResponse
     */
    'data'?: OccupationInfoResponse | null;
}